[TASK] Use automatic codesniffer ruleset detection
The package dealerdirect/phpcodesniffer-composer-installer depends on
codesniffer and automatically configures its installed_paths config setting. We
no longer have to overwrite that setting in order to get our ruleset to work.

This enables additional rulesets to be defined on the project level without
needing to think about when to set the installed_paths key.